Several changes
is_admin is now can_manage_channels. Added some error messages for when audit logs are send and the channel registered is now found in the guild. Now if you confess and the set channel is not found the bot will tell you so. Added on_guild_remove event which deletes the guild from the server dictionary.